fix: Fix homebrew path co-opting dev servers
Adding homebrew to global path puts it at end of a prefix block, but that
puts it ahead of changes made by shadowenv, which breaks servers in dev,
even if those servers don't use fish, they inherit the path with homebrew
at the start.

It is all very confusing and I don't totally understand it, but this definitely
fixes things in a pretty scalable way.

As a note to future self: using "set -U fish_user_paths" to totally erase the
universal variable is dangerous but very effective. Especially when it is set
by fish_add_path on the next run